+++ Code/LYRUIOutgoingMessageCollectionViewCell.m
@@ -14,7 +14,7 @@
 {
     self = [super initWithFrame:frame];
     if (self) {
-        self.avatarImage.hidden = YES;
+        self.avatarImageView.hidden = YES;
 
         [self.contentView addConstraint:[NSLayoutConstraint constraintWithItem:self.bubbleView
                                                                      attribute:NSLayoutAttributeRight