+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * `-[LYRUIMessageComposeTextView removeAttachements]` has been removed. It doesn't have a replacement since it was meant for internal use only.
 * The `placeHolderText` property of `LYRUIMessageComposeTextView` is now `placeholder`.
 * The unused `pendingBubbleViewColor` property of `LYRUIMessageCollectionViewCell` has been removed.
+* The `avatarImage` property of `LYRUIMessageCollectionViewCell` is now `avatarImageView`.
 * The `initialViewBackgroundColor` property of `LYRUIAvatarImageView` has been removed. Use `backgroundColor` instead.
 * The `initialFont` property of `LYRUIAvatarImageView` is now `initialsFont`.
 * The `initialColor` property of `LYRUIAvatarImageView` is now `initialsColor`.